Fix four extractor bugs that were hiding work from the sweeps, and start on SQL
Extending the sweeps to SQL meant reusing `eachTag`, and pointing it at a new
component surfaced defects that had been silently shrinking the *existing*
Python sweeps. Every one of them fails in the direction this module's own
header warns about: a component it cannot read looks like a component that
passed.

1. `eachTag` abandoned the whole file on an unterminated match. Frontmatter
   like `description: Sandbox for \`<SqlChallengeCard dialect="postgres">\``
   is not a tag and has no `/>`, so the generator hit the end of the file and
   `break`, discarding every real tag after it. Cost: ~84 `<ChallengeCard>`s
   across nine files and ~20 `<CodeBlock>`s across thirteen. Frontmatter is
   now skipped, and an unterminated tag in the body is yielded with
   `unterminated: true` so callers count it rather than lose it.

2. Prose about a component was matched as a use of it. "A variable defined in
   one `<CodeBlock>` is not visible to the next" appears in ten course
   overviews. Matches inside an inline code span are skipped.

   (A document-wide ``` fence parity count was tried and reverted: props carry
   fences of their own, so the count flips inside `markdown={…}` and starts
   discarding real tags after it. The comment in `insideCode` records why.)

3. `readTemplate` left `\n` as backslash-n. These props are template literals
   that MDX hands to JavaScript, where `\n` is a newline. A card authored on
   one line as `def f():\n    return 1` was extracted as Python containing a
   literal backslash and raised "unexpected character after line continuation
   character" — a parser bug reported as broken content. `\t` and `\r` too.
   The `"\\n"` a lesson prints is written `\\n` and still arrives intact.

4. Sibling modules leaked between cards. One interpreter serves the sweep but
   a reader gets a fresh runtime per card, so once any card's `utils.py` had
   been imported, `import utils` in a later card returned the first one and
   its function was missing. The failure landed on the innocent card, only in
   a full sweep, and never for a reader. `stage()` now drops the modules it is
   about to overwrite from `sys.modules`.

Extraction now agrees exactly with an independent count of the raw tags:
1696 python `<CodeBlock>`s and 316 `<ChallengeCard>`s, against 1694 and 306
before. `check:challenges` passes all 309 checkable cards (316 total, 7
environmental), up from 299.

Groundwork for the SQL sweep, which is the next commit:

  • `evaluateSqlTest` and the declarative test language it grades move out of
    SqlChallengeCard.tsx into `sqlChallengeHarness.ts`, so a script can run
    them. Same reasoning as `challengeHarness.ts`: a second copy of a grader
    drifts, and a drifted grader does not fail loudly, it passes cards it
    should have failed. The component re-exports the types, so no import
    elsewhere changes.
  • `extractSqlBlocks` / `extractSqlCards` read the 381 `<SqlCodeBlock>`s and
    73 `<SqlChallengeCard>`s (postgres 129/19, sqlite 129/7, duckdb 123/26).
  • PGlite and sqlite-wasm both boot and query correctly under Node, so those
    two dialects need no new dependency. duckdb-wasm is loaded from a CDN by
    the site and will need one pinned to `DUCKDB_VERSION`.
